OMT-246 add PremiumMutation for CreatePremium
This needs an update of the core module
